requirements.txt
boto3==1.4.4
dnspython==1.16.0
flake8==3.8.3
pymongo==3.10.1
python-dotenv==0.13.0
python-lambda==11.7.1
twilio==6.43.0
